[hotfix][table-planner-blink] Move function related tests to function package
