+++ flink-table/flink-table-planner-blink/src/test/java/org/apache/flink/table/planner/functions/ConstructedAccessFunctionITCase.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.expressions;
+package org.apache.flink.table.planner.functions;
 
 import org.apache.flink.table.annotation.DataTypeHint;
 import org.apache.flink.table.annotation.FunctionHint;
@@ -62,11 +62,11 @@ import static org.junit.Assert.assertThat;
 @RunWith(Suite.class)
 @Suite.SuiteClasses(
 	{
-		CompositeTypeAccessExpressionITCase.FieldAccessFromTable.class,
-		CompositeTypeAccessExpressionITCase.FieldAccessAfterCall.class
+		ConstructedAccessFunctionITCase.FieldAccessFromTable.class,
+		ConstructedAccessFunctionITCase.FieldAccessAfterCall.class
 	}
 )
-public class CompositeTypeAccessExpressionITCase {
+public class ConstructedAccessFunctionITCase {
 
 	/**
 	 * Regular tests. See also {@link FieldAccessAfterCall} for tests that access a nested field of an expression or
